style(seminar2): user 요청 관련 파일 페키징
