+++ Seminar2/GrowthAssignment/src/model/user/User.ts
